LiteLLM otel instrumentation is working
but it took installing all of litellm[proxy] ... worse, not all of it but most...
there is one conflicting dependency. So I had to install enough of
the proxy's dependencies to get it to load,
in order to provide its config or lack thereof to the initialization
of the instrumentation! Not OK!